crystal_bot_files/requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
cryptography==42.0.2
asyncio==3.4.3
typing==3.7.4.3